SL half

+++ OpenGLCourseApp/PointLight.cpp
@@ -29,7 +29,7 @@ PointLight::PointLight(GLfloat red, GLfloat green, GLfloat blue,
 
 void PointLight::UseLight(GLuint ambientIntensityLocation, GLuint ambientColourLocation,
               GLuint diffuseIntensityLocation, GLuint positionLocation,
-              GLfloat constantLocation, GLfloat linearLocation, GLfloat exponentLocation)
+              GLuint constantLocation, GLuint linearLocation, GLuint exponentLocation)
 {
     glUniform3f(ambientColourLocation, colour.x,colour.y,colour.z);
     glUniform1f(ambientIntensityLocation, ambientIntensity);
